requirements.txt
streamlit>=1.26.0
jq
langchain
langchain_openai
langchain-community
langchain-postgres
langchain-nvidia-ai-endpoints
fastembed
pgvector
psycopg[binary]
pypdf
python-dotenv
bs4
